Custom RUNIT ISO File (Custom Ubuntu)
I customized this version of ubuntu with C/C++ compilers, Python3 + Libraries, Java (JRE + JDK), and it also has custom preinstalled file directories with an autorun script that automatically runs code in a specificed directory and will send the output.txt file to an output folder directory. We will have to work on the code file injection next so we can automate code requests.
Since the file is too large to upload to github (6.8GB ISO), I uploaded a txt file with a link where you can download the ISO file.